basic-enemy-ai (#12)
Made the enemy way simpler.  When it sees you, it attacks.  Otherwise, it will wander around.

Keep in mind that what I wrote here is spaghetti code and most of the code in this enemy should be moved to the parent.  This is a quick fix, nothing solid and robust, but it works.